refactor: move duplicated deriveUserEvent and isUserEvent definitions to transaction utils file
